calendar-app/server/requirements.txt
Flask
google-api-python-client
google-auth-oauthlib
google-generativeai